pyproject.toml
[project]
name = "nano-banana-python"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "google-genai",
    "python-dotenv>=1.1.1",
    "fastapi[standard]>=0.115.0",
    "python-multipart>=0.0.9",
    "slowapi>=0.1.9",
]

[dependency-groups]
dev = ["pytest>=8.4.1"]
